Refactor: move CustomSaveStateHandle and implement it
