+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/shared/util/save_state/CustomSaveStateHandle.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.presentation.shared.util
+package net.thechance.mena.trends.presentation.shared.util.save_state
 
 interface CustomSaveStateHandle {
     fun <T> get(key:String) : T?
